added support for JsonReader already positioned fixed missing space in namespace
